(#428) code 钱包提现金额优化

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/topdynamic_comment/DynamicCommentTopContract.java
@@ -30,7 +30,7 @@ public interface DynamicCommentTopContract {
     }
 
     interface Presenter extends IBasePresenter {
-        float getBalance();
+        double getBalance();
         void topDynamicComment(long feed_id,long comment_id, int amount, int day);
     }
